python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.26.2
requests==2.31.0
insightface==0.7.3
onnxruntime==1.16.3
